Suggestions, Formatting, and cleanup
